# TypeScript: Reuse Reader for stream reader