… typescript/core/src/McapStreamReader.test.ts
… typescript/core/src/McapWriter.test.ts
… typescript/core/src/Reader.ts
… typescript/core/src/StreamBuffer.test.ts
… typescript/core/src/StreamBuffer.ts
… typescript/core/src/parse.ts